+++ src/Filter/Filter.fs
@@ -1,57 +1,71 @@
 namespace Lmc.KafkaApplication.Filter
 
+
+[<RequireQualifiedAccess>]
 module internal Filter =
-    open System.IO
-    open FSharp.Data
-    open Lmc.ServiceIdentification
     open Lmc.Consents.Intent
-    open Lmc.Kafka
-    open Lmc.KafkaApplication
+    open Lmc.ErrorHandling
+    open Lmc.ServiceIdentification
 
-    type private ConfigurationSchema = JsonProvider<"src/Filter/schema/configuration.json", SampleIsList=true>
+    [<RequireQualifiedAccess>]
+    module Configuration =
+        open FSharp.Data
+        open System.IO
 
-    let parseFilterConfiguration path =
-        let configuration =
-            path
-            |> File.ReadAllText
-            |> ConfigurationSchema.Parse
+        type private ConfigurationSchema = JsonProvider<"src/Filter/schema/configuration.json", SampleIsList=true>
 
-        {
-            Spots =
+        let parse path = result {
+            let configuration =
+                path
+                |> File.ReadAllText
+                |> ConfigurationSchema.Parse
+
+            let! spots =
                 configuration.Filter.Spot
-                |> Seq.map (fun spot -> { Zone = Zone spot.Zone; Bucket = Bucket spot.Bucket } )
+                |> Seq.map (fun parsedSpot -> Create.Spot(parsedSpot.Zone, parsedSpot.Bucket))
                 |> List.ofSeq
-            Intents =
-                configuration.Filter.Intents
-                |> Seq.map (fun intent -> { Purpose = IntentPurpose intent.Purpose; Scope = IntentScope intent.Scope } )
-                |> List.ofSeq
+                |> Validation.ofResults
+                |> Result.mapError InvalidSpot
+
+            return {
+                Spots = spots
+                Intents =
+                    configuration.Filter.Intents
+                    |> Seq.map (fun intent -> { Purpose = IntentPurpose intent.Purpose; Scope = IntentScope intent.Scope } )
+                    |> List.ofSeq
+            }
         }
 
-    let private isValueAllowed allowedValues = function
-        | None -> true
-        | Some value ->
-            match allowedValues with
-            | [] -> true
-            | allowedValues ->
-                allowedValues
-                |> List.contains value
+        let private isValueAllowed allowedValues = function
+            | None -> true
+            | Some value ->
+                match allowedValues with
+                | [] -> true
+                | allowedValues ->
+                    allowedValues
+                    |> List.contains value
 
-    let private isAllowedBy { Spots = spots; Intents = intents } (spot, intent) =
-        Some spot |> isValueAllowed spots
-        || intent |> isValueAllowed intents
+        let isAllowedBy { Spots = spots; Intents = intents } (spot, intent) =
+            Some spot |> isValueAllowed spots
+            || intent |> isValueAllowed intents
 
-    let filterByConfiguration getCommonEvent getIntent configuration tracedEvent =
-        let inputEvent = tracedEvent |> TracedEvent.event
+    [<RequireQualifiedAccess>]
+    module Filtering =
+        open Lmc.Kafka
+        open Lmc.KafkaApplication
 
-        let commonEvent: CommonEvent =
-            inputEvent
-            |> Input
-            |> getCommonEvent
-        let spot = { Zone = commonEvent.Zone; Bucket = commonEvent.Bucket}
+        let filterByConfiguration getCommonEvent getIntent configuration tracedEvent =
+            let inputEvent = tracedEvent |> TracedEvent.event
 
-        let intent: Intent option =
-            inputEvent
-            |> getIntent
+            let commonEvent: CommonEvent =
+                inputEvent
+                |> Input
+                |> getCommonEvent
+            let spot = Create.Spot(commonEvent.Zone, commonEvent.Bucket)
 
-        if (spot, intent) |> isAllowedBy configuration then Some tracedEvent
-        else None
+            let intent: Intent option =
+                inputEvent
+                |> getIntent
+
+            if (spot, intent) |> Configuration.isAllowedBy configuration then Some tracedEvent
+            else None