+++ src/Filter/Runner.fs
@@ -1,15 +1,16 @@
 namespace Lmc.KafkaApplication.Filter
 
 module internal FilterRunner =
+    open Microsoft.Extensions.Logging
     open Lmc.KafkaApplication
 
     let runFilter: RunPattern<FilterApplication<'InputEvent, 'OutputEvent>, 'InputEvent, 'OutputEvent> =
         fun run (FilterApplication application) ->
             let beforeRun filterApplication: BeforeRun<'InputEvent, 'OutputEvent> =
                 fun app ->
-                    filterApplication.FilterConfiguration
-                    |> sprintf "%A"
-                    |> app.Logger.VeryVerbose "Filter"
+                    app.LoggerFactory
+                        .CreateLogger("KafkaApplication.Filter")
+                        .LogDebug("Configuration: {configuration}", filterApplication.FilterConfiguration)
 
             application
             |> PatternRunner.runPattern (PatternName "Filter") FilterApplication.application beforeRun run