+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
 - Fix default spot from `(Zone=common, Bucket=all)` to `(Zone=all, Bucket=common)`
+- Update dependencies
+- [**BC**] Use `ILogger` instead of `ApplicationLogger`
+- [**BC**] Remove Graylog support
+    - Remove `logToGraylog`
+    - ...
+- [**BC**] Use `LoggerFactory` instead of `ApplicationLogger` and create other loggers by context
+    - Rename operation `useLogger` to `useLoggerFactory`
 
 ## 14.1.0 - 2021-09-20
 - Update dependencies